backend/requirements.txt
# Smart City Nashik — Python dependencies
# Compatible with Python 3.12, 3.13, 3.14

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Supabase client
supabase==2.9.1

# Environment variables
python-dotenv==1.0.1

# File uploads
python-multipart==0.0.12

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Image compression — use latest Pillow which supports Python 3.14
Pillow>=10.4.0

# HTTP client
httpx==0.27.2